+++ src/interfaces/database/item.ts
