+++ src/interfaces/database/usersingularity.ts
