+++ src/interfaces/singularity.ts
@@ -3,8 +3,8 @@ import { Model } from "mongoose";
 
 import { User } from "../database/schema/user";
 import { Server } from "../database/schema/server";
-import Command from "../database/interfaces/command";
-import Context from "../database/interfaces/context";
+import Command from "./client/command";
+import Context from "./client/context";
 
 export default interface Singularity extends Client {
 	userModel: Model<User>;
